+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## next version v0.0.12
 
--
+- variable `$id$` as part of ssid name in `softap.cfg` will be expanded to machine unique id
+- 
 
 
 ### backlog